GBValues files was created
